adding features for traits among other 3rd job related features

+++ models/area.go
@@ -9,6 +9,6 @@ type Area struct {
 }
 
 type LevelRange struct {
-	Max int64 `json:"max" firestore:"max"`
-	Min int64 `json:"min" firestore:"min"`
+	Max int32 `json:"max" firestore:"max"`
+	Min int32 `json:"min" firestore:"min"`
 }
