+++ models/level.go
@@ -1,6 +1,6 @@
 package models
 
 type Level struct {
-	Value int64 `json:"value" firestore:"value"`
+	Value int32 `json:"value" firestore:"value"`
 	Exp   int64 `json:"exp" firestore:"exp"`
 }
